upgrade node update dependencies rename detailsToggle
