+++ .nvmrc
@@ -1,1 +1,1 @@
-16.15.0
+18.16.0
