+++ app/client/javascript/details-toggle.ts
