+++ package.json
@@ -4,7 +4,7 @@
   "version": "0.0.1",
   "private": true,
   "engines": {
-    "node": ">=16.15.0"
+    "node": ">=18.16.0"
   },
   "scripts": {
     "start": "node dist/server/server",
@@ -16,7 +16,7 @@
     "build": "yarn build:prod && yarn transpile",
     "build:dev": "yarn build-sass && yarn build-js:dev",
     "build:prod": "yarn build-sass && yarn build-js",
-    "build-sass": "node-sass app/client/sass/application.scss public/css/main.css --output-style compressed",
+    "build-sass": "sass --load-path=./ --style=compressed app/client/sass/application.scss public/css/main.css",
     "build-js": "browserify ./app/client/javascript/main.ts -p [ tsify -p app/client ] | uglifyjs -c > ./public/js/main.js",
     "build-js:dev": "browserify ./app/client/javascript/main.ts -p [tsify -p app/client] --debug -o ./public/js/main.js ",
     "test": "yarn test:unit && yarn test:integration",
@@ -58,10 +58,10 @@
     "@types/express": "^4.17.14",
     "@types/express-serve-static-core": "^4.17.31",
     "applicationinsights": "^1.6.0",
-    "applicationinsights-native-metrics": "^0.0.8",
     "axios": "^0.27.0",
     "body-parser": "^1.20.1",
     "chokidar": "^3.5.3",
+    "cliui": "^8.0.1",
     "computed-style": "^0.3.0",
     "config": "3.3.9",
     "connect-redis": "^6.1.3",
@@ -70,9 +70,8 @@
     "eslint-plugin-mocha": "^10.1.0",
     "express": "^4.18.2",
     "express-session": "^1.17.3",
-    "govuk-elements-sass": "^3.1.3",
-    "govuk-frontend": "^4.3.1",
-    "govuk_template_jinja": "0.26.0",
+    "govuk-frontend": "^4.7.0",
+    "govuk_template_jinja": "^0.26.0",
     "helmet": "^6.0.0",
     "http-status-codes": "^2.0.0",
     "i": "^0.3.6",
@@ -87,13 +86,13 @@
     "moment": "^2.29.4",
     "multer": "^1.4.5-lts.1",
     "node-fetch": "^3.3.0",
-    "node-sass": "^8.0.0",
     "nunjucks": "^3.2.3",
     "otp": "^0.1.3",
     "promise-request-retry": "^1.0.2",
     "proxyquire": "^2.1.3",
     "reload": "^3.2.0",
     "request-promise": "^4.2.2",
+    "sass": "^1.66.1",
     "serve-favicon": "^2.4.2",
     "shelljs": "^0.8.5",
     "superagent": "^8.0.5"
@@ -112,7 +111,7 @@
     "@types/joi": "^13.5.2",
     "@types/jquery": "^3.5.14",
     "@types/mocha": "^10.0.0",
-    "@types/multer": "^1.4.7",
+    "@types/multer": "^1.4.8",
     "@types/node": "^16.11.7",
     "@types/nunjucks": "^3.1.3",
     "@types/shelljs": "^0.8.3",
@@ -127,7 +126,7 @@
     "chai-string": "^1.5.0",
     "cli-color": "^2.0.3",
     "codeceptjs": "^3.3.7",
-    "concurrently": "^7.0.0",
+    "concurrently": "^8.2.1",
     "cross-env": "^4.0.0",
     "domready": "^1.0.8",
     "dyson": "^4.1.0",
@@ -171,7 +170,8 @@
   "resolutions": {
     "json5": "^2.2.2",
     "http-cache-semantics": "^4.1.1",
-    "cookiejar": "^2.1.4"
+    "cookiejar": "^2.1.4",
+    "@types/glob": "^8.1.0"
   },
   "typeScriptVersion": "4.8",
   "packageManager": "yarn@3.6.0"
